backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
aiohttp
pydantic
sse-starlette
itsdangerous
gunicorn 